// src/app/user/saved/model/CreateInfoModel.ts
export interface CreateInfoModel{
    ruc: string;
    businessName: string;
    country: string;
    userId: number;
    stage: string;
    commentary: string;
    oppNumber: string;
    product: string;
    units: number;
    realRent: number;
    contact: string;
    contactNumber: string;
    email: string;
    expirationAt: string;
    updatedAt: string;
    closeAt: string;
    salesManagerId: number;
    executiveId: number;
}